+++ CHANGELOG.md
@@ -50,6 +50,8 @@
 - :sparkles: feat: config ViewResolver
 - :sparkles: feat: add ControllerAdvice,ExceptionHandler
 - :bug: fix: #7 filter can't be null
+- :sparkles: feat: use ControllerAdviceExceptionResolver as default ExceptionResolver
+
 
 
 ## v2.3.6